requirement.txt
alembic==1.7.3
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.10.0
bleach==4.1.0
bs4==0.0.1
certifi==2020.12.5
charset-normalizer==2.0.6
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
cycler==0.10.0
Django==3.2.14
django-ckeditor==6.3.2
django-cors-headers==3.10.1
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.14.0
docutils==0.17.1
dominate==2.6.0
drf-yasg==1.21.5
fake-useragent==0.1.11
Flask==1.0.3
flask-app==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
greenlet==1.1.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lightgbm==3.2.1
lxml==4.8.0
Mako==1.1.5
MarkupSafe==1.1.1
matplotlib==3.3.4
nltk==3.6.2
numpy==1.19.5
packaging==21.0
pandas==0.25.3
parse==1.19.0
patsy==0.5.1
Pillow==8.4.0
pkginfo==1.7.1
psycopg2-binary==2.8.3
pyee==8.2.2
Pygments==2.10.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2021.1
readme-renderer==29.0
regex==2021.7.6
reportlab==3.6.8
requests==2.26.0
requests-html==0.10.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==0.24.2
scipy==1.5.4
scorecardpy==0.1.9.3
six==1.15.0
soupsieve==2.3.2
SQLAlchemy==1.4.23
sqlparse==0.4.2
statsmodels==0.12.2
threadpoolctl==2.2.0
tqdm==4.61.2
twilio==7.17.0
twine==1.13.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.6
visitor==0.1.3
w3lib==1.22.0
webencodings==0.5.1
websockets==9.1
Werkzeug==1.0.1
wkhtmltopdf==0.2
zipp==3.5.0
